refactor: SnackReview 리팩토링 (#79)

+++ server/src/main/java/seb/project/Codetech/snackreview/entity/ReviewScore.java
@@ -26,7 +26,7 @@ public class ReviewScore {
 			}
 
 			try {
-				totalScore += field.getInt(this);
+				totalScore += field.getFloat(this);
 			} catch (IllegalAccessException e) {
 
 			} finally {
